+++ src/v1/HttpStatusCode/HttpStatusCode.ts
@@ -31,45 +31,16 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { ErrorTableTemplateWithNoExtraData } from "../ErrorTableTemplate";
-import { ExtraDataTemplate, NoExtraDataTemplate } from "../ExtraData";
-import { HttpStatusCode } from "../HttpStatusCode";
-import { PackageName } from "../PackageName";
-import { ErrorTable } from "./ErrorTable";
-import { UnreachableCodeTemplate } from "./UnreachableCode";
 
-// we can't use `packageNameFrom()` from `ts-lib-packagename` here, because
-// it creates a circular dependency that stops unit tests from compiling
-export const PACKAGE_NAME = "@ganbarodigital/ts-lib-error-reporting/lib/v1" as PackageName;
-
-/**
- * the ErrorTable for the package `@ganbarodigital/ts-lib-error-reporting`
- *
- * You can extend this in your own package, if you want to reuse the errors
- * that we provide.
- */
-export class PackageErrorTable implements ErrorTable {
-    // everything in this class has to follow the same structure
-    [key: string]: ErrorTableTemplateWithNoExtraData<any, string, ExtraDataTemplate | NoExtraDataTemplate>;
-
-    /**
-     * use this error in if/else & the default clause of switch statements
-     * to spot things that should never happen
-     */
-    public "unreachable-code": UnreachableCodeTemplate = {
-        packageName: PACKAGE_NAME,
-        errorName: "unreachable-code",
-        status: 500 as HttpStatusCode,
-        detail: "this code should never execute",
-        extra: {
-            logsOnly: {
-                reason: "explain why this code should be unreachable",
-            },
-        },
-    };
+// copied from ts-lib-value-objects, so that we can break the
+// circular dependency
+interface Branding<BrandingT extends string> {
+    BrandingT: BrandingT;
 }
 
+type Branded<T, BrandingT extends string> = T & Branding<BrandingT>;
+
 /**
- * a list of all of the errors that this package can throw
+ * represents any HTTP status code in the range `100` to `599` inclusive
  */
-export const ERROR_TABLE = new PackageErrorTable();
+export type HttpStatusCode = Branded<number, "@ganbarodigital/HttpStatusCode">;